Update types for redux store
